+++ gulp/fonts.js
@@ -1,7 +1,6 @@
 /* Fonts */
 var gulp = require('gulp'),
 		config = require('./config'),
-		newer = require('gulp-newer'),
 		gutil = require('gulp-util'),
 		browserSync = require("browser-sync"),
 		reload = browserSync.reload,
@@ -13,7 +12,6 @@ gulp.task('fonts', function() {
 			gutil.log(gutil.colors.red(error.message));
 			this.emit('end');
 		}))
-		.pipe(newer(config.pathTo.Build.Fonts))
 		.pipe(gulp.dest(config.pathTo.Build.Fonts))
 		.pipe(reload({stream: true}));
 });