+++ gulp/jade.js
@@ -1,13 +1,13 @@
 /* Jade */
 var gulp = require('gulp'),
-	config = require('./config'),
-	prettify = require('gulp-prettify'),
-	jade = require('gulp-jade'),
-	newer = require('gulp-newer'),
-	gutil = require('gulp-util'),
-	browserSync = require("browser-sync"),
-	reload = browserSync.reload,
-	plumber = require('gulp-plumber');
+		config = require('./config'),
+		prettify = require('gulp-prettify'),
+		jade = require('gulp-jade'),
+		changedInPlace = require('gulp-changed-in-place'),
+		gutil = require('gulp-util'),
+		browserSync = require("browser-sync"),
+		reload = browserSync.reload,
+		plumber = require('gulp-plumber');
 
 gulp.task('jade', function() {
 	return gulp.src(config.pathTo.Src.Jade)
@@ -15,7 +15,7 @@ gulp.task('jade', function() {
 			gutil.log(gutil.colors.red(error.message));
 			this.emit('end');
 		}))
-		.pipe(newer(config.pathTo.Build.Jade))
+		.pipe(changedInPlace())
 		.pipe(jade({
 			pretty: true
 		}))