+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#1.0
+* Fix all bugs
+* Change gulp-newer plugin on gulp-changer
+
 # 0.5
 * Fix templates for create page
 * Delete index default page